+++ conda/build-environment.yaml
@@ -1,4 +1,3 @@
-#!/bin/bash
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -16,16 +15,21 @@
 # specific language governing permissions and limitations
 # under the License.
 
-set -e
-set -u
+# Build environment.
+name: tvm-build
 
-source tests/scripts/setup-pytest-env.sh
+channels:
+  - anaconda
+  - conda-forge
 
-# cleanup pycache
-find . -type f -path "*.pyc" | xargs rm -f
-make cython3
-
-TVM_FFI=ctypes python3 -m pytest tests/python/all-platform-minimal-test
-TVM_FFI=cython python3 -m pytest tests/python/all-platform-minimal-test
-TVM_FFI=ctypes python3 -m pytest tests/python/unittest
-TVM_FFI=cython python3 -m pytest tests/python/unittest
+dependencies:
+  - conda-build
+  - git
+  - llvmdev ==10.0.0
+  - numpy
+  - pytest
+  - cython
+  - cmake
+  - bzip2
+  - make
+  - scipy
