+++ tests/python/all-platform-minimal-test/test_runtime_ndarray.py
@@ -14,6 +14,8 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+"""Basic runtime enablement test."""
+
 import tvm
 from tvm import te
 import numpy as np
